requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django>=3.2.13
django-cors-headers==3.12.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.4.0
python-environ==0.4.54
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.1.0
